# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FMD2r",
    "description": "",
    "license": {
      "name": "GPL-2.0-only",
      "identifier": "GPL-2.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/check-folders": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Check whether folders exist and are writable by the server (T74), for the destinations'\nwarnings. Relative paths resolve against the server's working directory, as downloads do.",
        "operationId": "checkFolders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "One check per path, in order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderCheck"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/modules/{id}/settings": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "A module's options, limits and overrides.",
        "operationId": "getModuleSettings",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Module ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The module's settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleSettingsView"
                }
              }
            }
          },
          "404": {
            "description": "No module with that ID is loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "modules"
        ],
        "summary": "Update a module's settings with a JSON merge patch (RFC 7396) over `enabled`, `limits`,\n`http`, `save_to` and `options` (option values keyed by `key`; `null` resets one to its default).\nNothing is stored unless the whole patch is valid.",
        "operationId": "patchModuleSettings",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Module ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleSettingsView"
                }
              }
            }
          },
          "400": {
            "description": "The patch is not a JSON object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No module with that ID is loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A value is invalid or a setting unknown; `field` names it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/preview-rename": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Preview the naming settings of a (possibly unsaved) draft on a sample chapter: the names\nand path the download engine gives it.",
        "operationId": "previewRename",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenamePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The expanded names",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenamePreview"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "All settings.",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Every setting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Update settings with a JSON merge patch (RFC 7396) over [`Settings`]; `null` resets a setting\nto its default. Nothing is stored unless the whole result is valid.",
        "operationId": "patchSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          },
          "400": {
            "description": "The patch is not a JSON object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A value is invalid or a setting unknown; `field` names it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/all": {
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Update the settings and any modules' settings together, all or nothing. Nothing is stored\nunless every patch is valid, and everything is stored in one transaction.",
        "operationId": "patchAllSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsSave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSettings"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No module with a given ID is loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Values are invalid or settings unknown; `fields` names each, prefixed `settings.` or `modules.<id>.`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FolderCheck": {
        "type": "object",
        "description": "Whether downloads can be saved in one folder.",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "problem": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why downloads can't be saved there now, or `None` when they can. A missing folder is\ncreated by the first download into it, so this is a warning, not an error: a disk may be\nunmounted for a while."
          }
        }
      },
      "FolderCheckRequest": {
        "type": "object",
        "description": "Folders to check, such as the destinations of a settings draft.",
        "required": [
          "paths"
        ],
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ModuleSettingsView": {
        "type": "object",
        "description": "A module's settings.",
        "required": [
          "id",
          "name",
          "enabled",
          "options",
          "limits",
          "module_limits",
          "http",
          "save_to"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Whether `limits` and `http` apply (`Settings.Enabled`); options always do."
          },
          "http": {
            "$ref": "#/components/schemas/HttpOverridesView"
          },
          "id": {
            "type": "string"
          },
          "limits": {
            "$ref": "#/components/schemas/LimitOverrides",
            "description": "The user's limit overrides."
          },
          "module_limits": {
            "$ref": "#/components/schemas/ModuleLimits",
            "description": "The limits the module itself declares; 0 means unlimited."
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleOptionSetting"
            },
            "description": "The options the module declares with `AddOption*`, in declaration order."
          },
          "save_to": {
            "type": "string",
            "description": "The folder the website's downloads go to when the user picks none; empty for the default\ndestination (`OverrideSettings.SaveToPath`, baseunits/WebsiteModulesSettings.pas:50).\nApplies whether or not `enabled` is set."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details body.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "The setting a validation error (422) is about, as a dotted path such as\n`connections.timeout_secs` or `options.server`.\nKept for clients that read one field: the first of `fields`."
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldProblem"
            },
            "description": "Every setting a validation error (422) is about, each with why it was rejected."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The status code's reason phrase."
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`: the status code says it all."
          }
        }
      },
...
      "RenamePreview": {
        "type": "object",
        "description": "The names a download of a sample chapter gets with a draft's settings.",
        "required": [
          "manga",
          "chapter",
          "filename",
          "page",
          "path"
        ],
        "properties": {
          "chapter": {
            "type": "string",
            "description": "The chapter folder or archive name (`chapter_rename`)."
          },
          "filename": {
            "type": "string",
            "description": "The first page's file name, without extension (`filename_rename`)."
          },
          "manga": {
            "type": "string",
            "description": "The manga folder name (`manga_rename`), whether or not the folder is generated."
          },
          "page": {
            "type": "string",
            "description": "The first page's file name with the extension it ends up with."
          },
          "path": {
            "type": "string",
            "description": "Where the first page ends up: its file, or the chapter's archive when chapters are\npacked."
          }
        }
      },
      "RenamePreviewRequest": {
        "type": "object",
        "description": "The settings a rename preview reads, possibly unsaved; a missing group takes its defaults.",
        "properties": {
          "images": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ImageSettings"
              }
            ],
            "default": {
              "imagemagick": {
                "compression": "None",
                "enabled": false,
                "quality": 75,
                "save_as": "JPEG"
              },
              "jpeg_quality": 80,
              "png_compression": "fastest",
              "png_to_jpeg": false,
              "webp_save_as": "png"
            }
          },
          "output": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OutputSettings"
              }
            ],
            "default": {
              "format": "folder",
              "pdf_quality": 100
            }
          },
          "saveto": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SaveToSettings"
              }
            ],
            "default": {
              "chapter_rename": "%CHAPTER%",
              "convert_digit_chapter": true,
              "convert_digit_volume": true,
              "default_dir": "downloads",
              "destinations": [
                {
                  "default": true,
                  "name": "Downloads",
                  "path": "downloads"
                }
              ],
              "digit_chapter_length": 3,
              "digit_volume_length": 2,
              "filename_rename": "%FILENAME%",
              "generate_chapter_folder": true,
              "generate_manga_folder": true,
              "illegal_chars": "posix",
              "manga_rename": "%MANGA%",
              "remove_manga_name_from_chapter": false,
              "replace_unicode": false,
              "replace_unicode_with": "_"
            }
          }
        }
      },
...
      "SavedSettings": {
        "type": "object",
        "description": "What [`patch_all`] saved.",
        "required": [
          "settings",
          "modules"
        ],
        "properties": {
          "modules": {
            "type": "object",
            "description": "The settings of each patched module, by module ID.",
            "additionalProperties": {
              "$ref": "#/components/schemas/ModuleSettingsView"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsView"
          }
        }
      },
...
      "SettingsSave": {
        "type": "object",
        "description": "The body of [`patch_all`]; either part may be left out.",
        "properties": {
          "modules": {
            "type": [
              "object",
              "null"
            ],
            "description": "A merge patch as for `PATCH /api/modules/{id}/settings`, by module ID.",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {},
              "propertyNames": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "settings": {
            "type": [
              "object",
              "null"
            ],
            "description": "A merge patch as for `PATCH /api/settings`.",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "SettingsView": {
        "type": "object",
        "description": "Every application setting, with the secrets left out: [`Settings`] as the API shows it.\nEach group's schema carries its defaults.",
        "properties": {
          "connections": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectionSettingsView"
              }
            ],
            "default": {
              "always_start_from_failed_chapters": true,
              "auto_retry_failed_tasks": 1,
              "flaresolverr_url": "",
              "max_favorite_threads": 1,
              "max_parallel_tasks": 1,
              "max_update_list_threads": 1,
              "proxy": {
                "enabled": false,
                "has_password": false,
                "host": "",
                "port": null,
                "type": "http",
                "username": ""
              },
              "retry_count": 5,
              "threads_per_task": 1,
              "timeout_secs": 30,
              "user_agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/133.0.0.0 Safari/537.36"
            }
          },
          "covers": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CoverSettings"
              }
            ],
            "default": {
              "cache_size_mb": 256,
              "revalidate_after_hours": 168
            }
          },
          "favorites": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FavoriteSettings"
              }
            ],
            "default": {
              "auto_download": false,
              "check_at_startup": true,
              "check_interval_minutes": 60,
              "check_on_interval": true,
              "remove_completed": false
            }
          },
          "general": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GeneralSettings"
              }
            ],
            "default": {
              "add_as_stopped": false,
              "data_dir": "data",
              "language": "en",
              "load_covers": true,
              "lua_dir": "lua",
              "selected_websites": []
            }
          },
          "images": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ImageSettings"
              }
            ],
            "default": {
              "imagemagick": {
                "compression": "None",
                "enabled": false,
                "quality": 75,
                "save_as": "JPEG"
              },
              "jpeg_quality": 80,
              "png_compression": "fastest",
              "png_to_jpeg": false,
              "webp_save_as": "png"
            }
          },
          "logs": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LogSettings"
              }
            ],
            "default": {
              "max_file_size_mb": 10,
              "max_files": 5
            }
          },
          "module_updater": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ModuleUpdaterSettingsView"
              }
            ],
            "default": {
              "auto_update": true,
              "has_github_token": false,
              "interval_minutes": 60,
              "keep_last_good": true,
              "repo_name": "FMD2",
              "repo_owner": "dazedcat19",
              "repo_path": "lua",
              "repo_ref": "master"
            }
          },
          "output": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OutputSettings"
              }
            ],
            "default": {
              "format": "folder",
              "pdf_quality": 100
            }
          },
          "saveto": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SaveToSettings"
              }
            ],
            "default": {
              "chapter_rename": "%CHAPTER%",
              "convert_digit_chapter": true,
              "convert_digit_volume": true,
              "default_dir": "downloads",
              "destinations": [
                {
                  "default": true,
                  "name": "Downloads",
                  "path": "downloads"
                }
              ],
              "digit_chapter_length": 3,
              "digit_volume_length": 2,
              "filename_rename": "%FILENAME%",
              "generate_chapter_folder": true,
              "generate_manga_folder": true,
              "illegal_chars": "posix",
              "manga_rename": "%MANGA%",
              "remove_manga_name_from_chapter": false,
              "replace_unicode": false,
              "replace_unicode_with": "_"
            }
          },
          "server": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ServerSettingsView"
              }
            ],
            "default": {
              "bind": "127.0.0.1:8080",
              "has_auth_token": false,
              "session_idle_days": 7,
              "session_lifetime_days": 30
            }
          },
          "update_lists": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UpdateListSettings"
              }
            ],
            "default": {
              "auto_update": false,
              "db_url": "https://raw.githubusercontent.com/dazedcat19/FMD2-DB/master/7z/<website>.7z",
              "interval_hours": 24,
              "new_manga_days": 1,
              "no_manga_info": false,
              "remove_duplicate_local_data": false
            }
          },
          "xpath": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/XPathSettings"
              }
            ],
            "default": {
              "backend": "native"
            }
          }
        }
      },
...
    }
  }
}